[ADD] l10n_ch_isr_payment_grouping
Definition a specific grouping rule for ISR payments

Swiss ISR payments must not be grouped as we want to keep a single transaction
per reference. The ISR payment reference doesn't imply additionnal fees from
the bank.

One transaction per ISR reference is needed in Swiss SEPA payments
to ensure a end-to-end flow that will ease the reconciliation on
the other end.

When ISR reference is detected we also avoid to concatenate the references
to not concatenate the same reference multiple times. In case the
same reference is sent by the supplier on different dates.